First steps supporting cells format
