+++ pkg/input/congolway_test.go
@@ -133,7 +133,7 @@ func TestNewGolFromTextFile10x10BadVersion(t *testing.T) {
 	dataFilePath := path.Join(currentDir, "..", "..", "testdata", "10x10_bad_version.txt")
 
 	gr := NewGolReader(new(gol.Gol))
-	g, error := gr.ReadGolFromTextFile(dataFilePath)
+	g, error := gr.ReadCongolwayFile(dataFilePath)
 
 	expectedError := "Unknonwn version found 999999"
 
@@ -150,71 +150,25 @@ func TestNewGolFromTextFile10x10BadVersion(t *testing.T) {
 	}
 }
 
-func testNewGolFromTextFile(t *testing.T, filename string,
-	rows int, cols int, limitRows bool, limitCols bool,
+func testNewGolFromTextFile(t *testing.T, filename string, rows int, cols int, limitRows bool, limitCols bool,
 	generation int, expectedCells [][]int) {
 
-	g, error := readGolFromTextFile(filename)
+	g, error := readCongolwayFile(filename)
 	if error != nil {
 		t.Error(error)
 		return
 	}
-
-	if g.Generation() != generation {
-		t.Errorf("Loaded generation is wrong, got: %d, must be: %d.", g.Generation(), generation)
-		return
-	}
-
-	if g.NeighborhoodTypeString() != "Moore" {
-		t.Errorf("Loaded neighborhood is wrong, got: %s, must be: Moore.", g.NeighborhoodTypeString())
-		return
-	}
-
-	if g.Rows() != rows {
-		t.Errorf("Loaded number of rows is wrong, got: %d, must be: %d.", g.Rows(), rows)
-		return
-	}
-	if g.Cols() != cols {
-		t.Errorf("Loaded number of cols is wrong, got: %d, must be: %d.", g.Cols(), cols)
-		return
-	}
-
-	if g.LimitRows() != limitRows {
-		if limitRows {
-			t.Errorf("Should limit rows, but it isn't.")
-		} else {
-			t.Errorf("Shouldn't limit rows, but it is.")
-		}
-		return
-	}
-	if g.LimitCols() != limitCols {
-		if limitCols {
-			t.Errorf("Should limit cols, but it isn't.")
-		} else {
-			t.Errorf("Shouldn't limit cols, but it is.")
-		}
-		return
-	}
-
-	for i := 0; i < rows; i++ {
-		for j := 0; j < cols; j++ {
-			gIJ := g.Get(i, j)
-			expectedIJ := expectedCells[i][j]
-			if gIJ != expectedIJ {
-				t.Errorf("Invalid cell at %d,%d. It got: %d, must be: %d.", i, j, gIJ, expectedIJ)
-			}
-		}
-	}
+	assertGolIsRight(t, filename, rows, cols, limitRows, limitCols, generation, expectedCells, g)
 }
 
-func readGolFromTextFile(filename string) (base.GolInterface, error) {
+func readCongolwayFile(filename string) (base.GolInterface, error) {
 	dataFilePath, dataFilePathError := base.GetTestdataFilePath(filename)
 	if dataFilePathError != nil {
 		return nil, dataFilePathError
 	}
 
 	gr := NewGolReader(new(gol.Gol))
-	gol, golReadError := gr.ReadGolFromTextFile(dataFilePath)
+	gol, golReadError := gr.ReadCongolwayFile(dataFilePath)
 	if golReadError != nil {
 		return nil, fmt.Errorf("Couldn't load the file %s: %s", dataFilePath, golReadError)
 	}
