+++ testdata/5x5.cells
@@ -1,0 +1,7 @@
+!Name: A 5x5 game of life grid
+!
+.....
+OOOOO
+OO.OO
+OOOOO
+.....
